requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# BERT and NLP
sentence-transformers>=2.2.0
nltk>=3.8.0
spacy>=3.5.0

# Explainability
shap>=0.42.0

# Fairness
fairlearn>=0.9.0
aif360>=0.5.0

# Visualization and Dashboard
streamlit>=1.25.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
datasets>=2.12.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
